+++ frontend/app/(auth)/logout/page.tsx
@@ -1,8 +1,8 @@
 "use client";
-import Button from "@/app/components/ui/Button";
-import Card from "@/app/components/ui/Card";
-import PageHeading from "@/app/components/ui/PageHeading";
 import { useSupabase } from "@/app/supabase-provider";
+import Button from "@/lib/components/ui/Button";
+import Card from "@/lib/components/ui/Card";
+import PageHeading from "@/lib/components/ui/PageHeading";
 import { useToast } from "@/lib/hooks/useToast";
 import Link from "next/link";
 import { useRouter } from "next/navigation";
