+++ frontend/app/chat/[chatId]/page.tsx
@@ -1,7 +1,7 @@
 "use client";
+import PageHeading from "@/lib/components/ui/PageHeading";
 import { UUID } from "crypto";
 import { useEffect } from "react";
-import PageHeading from "../../components/ui/PageHeading";
 import useChatsContext from "../ChatsProvider/hooks/useChatsContext";
 import { ChatInput, ChatMessages } from "../components";
