+++ frontend/app/chat/components/ChatMessages/ChatInput/ConfigButton.tsx
@@ -1,7 +1,7 @@
 "use client";
+import Button from "@/lib/components/ui/Button";
 import Link from "next/link";
 import { MdSettings } from "react-icons/md";
-import Button from "../../../../components/ui/Button";
 
 export function ConfigButton() {
   return (
