+++ frontend/app/config/components/UserAccountSection.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import { useSupabase } from "@/app/supabase-provider";
+import Button from "@/lib/components/ui/Button";
 import Link from "next/link";
-import Button from "../../components/ui/Button";
 
 export const UserAccountSection = (): JSX.Element => {
   const { session } = useSupabase();
