+++ frontend/app/upload/components/FileUploader/index.tsx
@@ -1,7 +1,7 @@
 "use client";
+import Button from "@/lib/components/ui/Button";
+import Card from "@/lib/components/ui/Card";
 import { AnimatePresence } from "framer-motion";
-import Button from "../../../components/ui/Button";
-import Card from "../../../components/ui/Card";
 import FileComponent from "./components/FileComponent";
 import { useFileUploader } from "./hooks/useFileUploader";
