+++ frontend/app/upload/page.tsx
@@ -1,8 +1,8 @@
 "use client";
+import Button from "@/lib/components/ui/Button";
+import { Divider } from "@/lib/components/ui/Divider";
+import PageHeading from "@/lib/components/ui/PageHeading";
 import Link from "next/link";
-import Button from "../components/ui/Button";
-import { Divider } from "../components/ui/Divider";
-import PageHeading from "../components/ui/PageHeading";
 import { Crawler } from "./components/Crawler";
 import { FileUploader } from "./components/FileUploader";
