+++ frontend/app/user/page.tsx
@@ -1,8 +1,8 @@
 "use client";
+import Spinner from "@/lib/components/ui/Spinner";
 import { useAxios } from "@/lib/useAxios";
 import { redirect } from "next/navigation";
 import { useEffect, useState } from "react";
-import Spinner from "../components/ui/Spinner";
 import { useSupabase } from "../supabase-provider";
 import { UserStatistics } from "./components/UserStatistics";
 import { UserStats } from "./types";
