+++ frontend/lib/components/ui/Toast/domain/types.ts
