+++ frontend/lib/hooks/useToast.ts
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { ToastContext } from "../../app/components/ui/Toast/domain/ToastContext";
+import { ToastContext } from "../components/ui/Toast/domain/ToastContext";
 
 export const useToast = () => {
   const { publish } = useContext(ToastContext);
